refactor(video): split videoController into focused modules
Extract the helper layers of the 1268-line videoController.ts into
controllers/video/: uploadMiddleware (multer config), mediaProbe
(ffprobe/ffmpeg helpers), uploadHelpers (upload payload pipeline +
result builders), channelUrl (YouTube/Bilibili/Twitch channel
resolution), mountVideo (mount path validation + headers), visibility
(role scoping), responses, and pathUtils.

videoController.ts stays the route-facing facade: it keeps the 11
request handlers and re-exports the upload middleware, so routes and
tests (which only import the public handlers + middleware) are
unaffected. Behavior is unchanged; the entry file drops 1268 -> 599 lines.
