+++ backend/src/controllers/videoController.ts
@@ -2,544 +2,58 @@
 import crypto from "crypto";
 import { Request, Response } from "express";
 import fs from "fs-extra";
-import multer from "multer";
 import path from "path";
-import sanitizeFilename from "sanitize-filename";
-import { IMAGES_DIR, SUBTITLES_DIR, VIDEOS_DIR } from "../config/paths";
+import { SUBTITLES_DIR, VIDEOS_DIR } from "../config/paths";
 import { NotFoundError, ValidationError } from "../errors/DownloadErrors";
 import { syncMediaServerArtifactsForRecord } from "../services/mediaServerExport";
-import { isLoginRequired } from "../services/passwordService";
 import * as storageService from "../services/storageService";
-import {
-  deleteSmallThumbnailMirrorSync,
-  regenerateSmallThumbnailForThumbnailPath,
-  resolveManagedThumbnailTarget,
-} from "../services/thumbnailMirrorService";
-import { twitchApiService } from "../services/twitchService";
-import {
-  extractTwitchVideoId,
-  isBilibiliUrl,
-  isTwitchChannelUrl,
-  isTwitchVideoUrl,
-  isYouTubeUrl,
-  normalizeTwitchChannelUrl,
-} from "../utils/helpers";
 import { logger } from "../utils/logger";
 import { getStringParam } from "../utils/paramUtils";
 import { successResponse } from "../utils/response";
-import {
-  createUploadValidationError,
-  createVideoUploadStorage,
-  getUploadVideoId,
-  UploadedVideoFile,
-} from "../utils/videoUpload";
+import { createUploadValidationError } from "../utils/videoUpload";
 import { resolvePlayableVideoFilePath } from "../utils/videoFileResolver";
 import {
   createReadStreamSafe,
   createWriteStreamSafe,
   pathExistsSafeSync,
-  pathExistsTrustedSync,
-  normalizeSafeAbsolutePath,
   resolveSafePath,
   sanitizePathSegment,
-  statTrustedSync,
   unlinkSafeSync,
   writeFileSafeSync,
 } from "../utils/security";
+import { getVisibilityScopedRole } from "./video/visibility";
+import { sendData, sendSuccess } from "./video/responses";
+import { getLanguageFromFilename, resolveVideoWebPath } from "./video/pathUtils";
+import {
+  buildUploadFailureResult,
+  buildUploadResponseMessage,
+  getDefaultUploadTitle,
+  getUploadSummary,
+  getUploadVideoPayload,
+  getUploadedFilesFromRequest,
+  readStringField,
+  UploadResultItem,
+} from "./video/uploadHelpers";
+import {
+  getExistingVideoBySourceUrl,
+  persistChannelUrlForVideo,
+  resolveChannelUrl,
+} from "./video/channelUrl";
+import {
+  assertMountFileExists,
+  isMountVideoPath,
+  resolveMountFilePath,
+  setMountVideoHeaders,
+} from "./video/mountVideo";
 
-/**
- * Resolve the caller role to pass into the visibility-aware storage queries
- * (`getVideos`/`getVideoById`). When login is disabled (single-user mode) every
- * caller is owner-equivalent, so any stale visitor session/JWT role left on the
- * request must be ignored — otherwise hidden videos would wrongly 404 / vanish
- * right after login is turned off. Only scope by role while login is enforced.
- */
-const getVisibilityScopedRole = (
-  req: Request
-):
-  | import("../services/storageService").VideoCallerRole
-  | undefined => {
-  if (!isLoginRequired()) {
-    return undefined;
-  }
-  return req.user?.role as
-    | import("../services/storageService").VideoCallerRole
-    | undefined;
-};
-
-const MAX_VIDEO_UPLOAD_FILE_SIZE = 100 * 1024 * 1024 * 1024;
-const MAX_BATCH_UPLOAD_FILES = 100;
-const MAX_BATCH_UPLOAD_TOTAL_SIZE = MAX_VIDEO_UPLOAD_FILE_SIZE;
-const MAX_SINGLE_UPLOAD_FIELDS = 4;
-const MAX_BATCH_UPLOAD_FIELDS = MAX_BATCH_UPLOAD_FILES + 4;
-
-export const videoUploadStorage = createVideoUploadStorage(VIDEOS_DIR);
-export const videoBatchUploadStorage = createVideoUploadStorage(VIDEOS_DIR, {
-  maxTotalBytes: MAX_BATCH_UPLOAD_TOTAL_SIZE,
-});
-
-const videoUploadOptions: multer.Options = {
-  storage: videoUploadStorage,
-  limits: {
-    fileSize: MAX_VIDEO_UPLOAD_FILE_SIZE,
-    files: 1,
-    fields: MAX_SINGLE_UPLOAD_FIELDS,
-    parts: 1 + MAX_SINGLE_UPLOAD_FIELDS,
-  },
-};
-
-const videoBatchUploadOptions: multer.Options = {
-  storage: videoBatchUploadStorage,
-  limits: {
-    fileSize: MAX_VIDEO_UPLOAD_FILE_SIZE,
-    files: MAX_BATCH_UPLOAD_FILES,
-    fields: MAX_BATCH_UPLOAD_FIELDS,
-    parts: MAX_BATCH_UPLOAD_FILES + MAX_BATCH_UPLOAD_FIELDS,
-  },
-};
-
-export const upload = multer(videoUploadOptions);
-export const uploadBatch = multer(videoBatchUploadOptions);
-
-export const uploadSubtitleMiddleware = multer({
-  storage: multer.memoryStorage(),
-  limits: {
-    fileSize: 10 * 1024 * 1024, // 10MB
-  },
-  fileFilter: (_req, file, cb) => {
-    if (file.originalname.match(/\.(vtt|srt|ass|ssa)$/i)) {
-      cb(null, true);
-    } else {
-      cb(
-        new Error(
-          "Invalid file type. Only .vtt, .srt, .ass and .ssa are allowed."
-        )
-      );
-    }
-  },
-});
-
-// Helper to safely execute files (prevent command injection)
-const execFileSafe = async (
-  file: string,
-  args: string[],
-  options: any = {}
-): Promise<void> => {
-  const { execFile } = await import("child_process");
-  return new Promise((resolve, reject) => {
-    execFile(file, args, options, (error, _stdout, stderr) => {
-      if (error) {
-        logger.error(`execFile error: ${error.message}`);
-        logger.error(`stderr: ${stderr}`);
-        reject(error);
-        return;
-      }
-      resolve();
-    });
-  });
-};
-
-// Helper validate paths functions
-const validateVideoPath = (inputPath: string): string => {
-  try {
-    return resolveSafePath(inputPath, VIDEOS_DIR);
-  } catch {
-    throw new ValidationError("Invalid video path", "path");
-  }
-};
-
-const getStoredUploadFilename = (
-  file: Pick<UploadedVideoFile, "filename">,
-): string | undefined => {
-  if (typeof file.filename !== "string" || file.filename.length === 0) {
-    return undefined;
-  }
-
-  const safeFilename = sanitizeFilename(file.filename);
-  if (!safeFilename || safeFilename !== file.filename) {
-    return undefined;
-  }
-
-  return safeFilename;
-};
-
-// Helper for video duration
-const getVideoDuration = async (
-  videoPath: string
-): Promise<number | undefined> => {
-  try {
-    const { execFile } = await import("child_process");
-    return new Promise((resolve) => {
-      execFile(
-        "ffprobe",
-        [
-          "-v",
-          "error",
-          "-show_entries",
-          "format=duration",
-          "-of",
-          "default=noprint_wrappers=1:nokey=1",
-          videoPath,
-        ],
-        (error, stdout) => {
-          if (error) {
-            logger.error("ffprobe error:", error);
-            resolve(undefined);
-            return;
-          }
-          const duration = parseFloat(stdout);
-          resolve(isNaN(duration) ? undefined : duration);
-        }
-      );
-    });
-  } catch (err) {
-    logger.error("Failed to get video duration:", err);
-    return undefined;
-  }
-};
-
-const uploadedFileHasVideoStream = async (videoPath: string): Promise<boolean> => {
-  try {
-    const { execFile } = await import("child_process");
-    return new Promise((resolve) => {
-      execFile(
-        "ffprobe",
-        [
-          "-v",
-          "error",
-          "-show_entries",
-          "stream=codec_type",
-          "-of",
-          "default=noprint_wrappers=1:nokey=1",
-          videoPath,
-        ],
-        (error, stdout) => {
-          if (error) {
-            logger.error("ffprobe stream validation error:", error);
-            resolve(false);
-            return;
-          }
-
-          const streamTypes = stdout
-            .split(/\r?\n/)
-            .map((line) => line.trim().toLowerCase())
-            .filter(Boolean);
-          resolve(streamTypes.includes("video"));
-        }
-      );
-    });
-  } catch (err) {
-    logger.error("Failed to validate uploaded video stream:", err);
-    return false;
-  }
-};
-
-const cleanupUploadedVideo = (
-  file: Pick<UploadedVideoFile, "filename"> | undefined,
-): void => {
-  const storedVideoFilename = file ? getStoredUploadFilename(file) : undefined;
-  if (!storedVideoFilename) {
-    return;
-  }
-
-  try {
-    const storedVideoPath = validateVideoPath(
-      `${VIDEOS_DIR}/${storedVideoFilename}`,
-    );
-    if (pathExistsSafeSync(storedVideoPath, VIDEOS_DIR)) {
-      unlinkSafeSync(storedVideoPath, VIDEOS_DIR);
-    }
-  } catch (error) {
-    logger.error("Failed to clean up uploaded video:", error);
-  }
-};
-
-const cleanupGeneratedThumbnail = (thumbnailPath: string | undefined): void => {
-  if (!thumbnailPath) {
-    return;
-  }
-
-  try {
-    if (pathExistsSafeSync(thumbnailPath, [IMAGES_DIR, VIDEOS_DIR])) {
-      unlinkSafeSync(thumbnailPath, [IMAGES_DIR, VIDEOS_DIR]);
-    }
-    deleteSmallThumbnailMirrorSync(thumbnailPath);
-  } catch (error) {
-    logger.error("Failed to clean up generated thumbnail:", error);
-  }
-};
-
-type UploadStatus = "uploaded" | "duplicate" | "failed";
-
-interface UploadResultItem {
-  originalName: string;
-  status: UploadStatus;
-  message: string;
-  video?: import("../services/storageService").Video;
-}
-
-interface UploadSummary {
-  total: number;
-  uploaded: number;
-  duplicates: number;
-  failed: number;
-}
-
-const getUploadSummary = (results: UploadResultItem[]): UploadSummary =>
-  results.reduce(
-    (summary, result) => {
-      summary.total += 1;
-      if (result.status === "uploaded") {
-        summary.uploaded += 1;
-      } else if (result.status === "duplicate") {
-        summary.duplicates += 1;
-      } else {
-        summary.failed += 1;
-      }
-      return summary;
-    },
-    { total: 0, uploaded: 0, duplicates: 0, failed: 0 }
-  );
-
-const getDefaultUploadTitle = (originalName: string): string =>
-  path.basename(originalName).replace(/\.[^/.]+$/, "");
-
-const readStringField = (
-  value: unknown,
-  fallback: string = ""
-): string => {
-  if (typeof value === "string") {
-    return value.trim() || fallback;
-  }
-  if (Array.isArray(value)) {
-    const firstString = value.find(
-      (entry): entry is string => typeof entry === "string" && entry.trim().length > 0
-    );
-    return firstString?.trim() || fallback;
-  }
-  return fallback;
-};
-
-const buildUploadResponseMessage = (summary: UploadSummary): string => {
-  if (summary.failed === 0 && summary.duplicates === 0) {
-    return `Uploaded ${summary.uploaded} video${summary.uploaded === 1 ? "" : "s"} successfully.`;
-  }
-
-  return `Upload finished. Uploaded: ${summary.uploaded}, duplicates skipped: ${summary.duplicates}, failed: ${summary.failed}.`;
-};
-
-const buildUploadFailureResult = (
-  file: Pick<UploadedVideoFile, "originalname">,
-  message: string
-): UploadResultItem => ({
-  originalName: path.basename(file.originalname),
-  status: "failed",
-  message,
-});
-
-const buildUploadDuplicateResult = (
-  file: Pick<UploadedVideoFile, "originalname">,
-  video: import("../services/storageService").Video
-): UploadResultItem => ({
-  originalName: path.basename(file.originalname),
-  status: "duplicate",
-  message: "Video already exists. Skipped duplicate upload",
-  video,
-});
-
-const buildUploadSuccessResult = (
-  file: Pick<UploadedVideoFile, "originalname">,
-  video: import("../services/storageService").Video,
-  status: Exclude<UploadStatus, "failed"> = "uploaded"
-): UploadResultItem => ({
-  originalName: path.basename(file.originalname),
-  status,
-  message:
-    status === "uploaded"
-      ? "Video uploaded successfully"
-      : "Video already exists. Skipped duplicate upload",
-  video,
-});
-
-const getUploadedFilesFromRequest = (req: Request): UploadedVideoFile[] => {
-  if (Array.isArray(req.files)) {
-    return req.files as UploadedVideoFile[];
-  }
-
-  if (req.file) {
-    return [req.file as UploadedVideoFile];
-  }
-
-  const filesByField = req.files as Record<string, Express.Multer.File[]> | undefined;
-  if (!filesByField) {
-    return [];
-  }
-
-  return Object.values(filesByField).flat() as UploadedVideoFile[];
-};
-
-const getUploadVideoPayload = async (
-  file: UploadedVideoFile,
-  title: string,
-  author: string
-): Promise<UploadResultItem> => {
-  const storedVideoFilename = getStoredUploadFilename(file);
-
-  if (file.validationError) {
-    cleanupUploadedVideo(file);
-    return buildUploadFailureResult(file, file.validationError);
-  }
-
-  if (!storedVideoFilename) {
-    return buildUploadFailureResult(file, "No video file uploaded");
-  }
-
-  // nosemgrep: javascript.lang.security.audit.path-traversal.path-join-resolve-traversal.path-join-resolve-traversal
-  const rawVideoPath = path.join(VIDEOS_DIR, storedVideoFilename);
-  const validatedVideoPath = validateVideoPath(rawVideoPath);
-  const contentHash = file.contentHash;
-  const videoId =
-    typeof contentHash === "string" && contentHash.length > 0
-      ? getUploadVideoId(contentHash)
-      : crypto.randomUUID();
-  const existingVideo = storageService.getVideoById(videoId);
-
-  if (existingVideo) {
-    cleanupUploadedVideo(file);
-    return buildUploadDuplicateResult(file, existingVideo);
-  }
-
-  const videoFilename = storedVideoFilename;
-  const requestedThumbnailFilename = `${path.parse(videoFilename).name}.jpg`;
-  const settings = storageService.getSettings();
-  const thumbnailTarget = resolveManagedThumbnailTarget(
-    {
-      videoPath: `/videos/${videoFilename}`,
-    },
-    requestedThumbnailFilename,
-    settings.moveThumbnailsToVideoFolder || false,
-  );
-  const thumbnailFilename = path.basename(thumbnailTarget.relativePath);
-  const validatedThumbnailPath = thumbnailTarget.absolutePath;
-  const thumbnailWebPath = thumbnailTarget.webPath;
-
-  const hasVideoStream = await uploadedFileHasVideoStream(validatedVideoPath);
-  if (!hasVideoStream) {
-    cleanupUploadedVideo(file);
-    cleanupGeneratedThumbnail(validatedThumbnailPath);
-    return buildUploadFailureResult(
-      file,
-      "Uploaded file is not a valid supported video"
-    );
-  }
-
-  try {
-    await execFileSafe("ffmpeg", [
-      "-i",
-      validatedVideoPath,
-      "-ss",
-      "00:00:00",
-      "-vframes",
-      "1",
-      validatedThumbnailPath,
-    ]);
-  } catch (error) {
-    logger.error("Error generating thumbnail:", error);
-  }
-
-  const thumbnailExists = pathExistsSafeSync(validatedThumbnailPath, [
-    IMAGES_DIR,
-    VIDEOS_DIR,
-  ]);
-  if (thumbnailExists) {
-    await regenerateSmallThumbnailForThumbnailPath(thumbnailWebPath);
-  }
-
-  const duration = await getVideoDuration(validatedVideoPath);
-
-  const fileSize =
-    typeof file.size === "number" && Number.isFinite(file.size)
-      ? file.size.toString()
-      : undefined;
-
-  const newVideo = {
-    id: videoId,
-    title: title || getDefaultUploadTitle(file.originalname),
-    author: author || "Admin",
-    source: "local",
-    sourceUrl: "",
-    videoFilename,
-    thumbnailFilename: thumbnailExists ? thumbnailFilename : undefined,
-    videoPath: `/videos/${videoFilename}`,
-    thumbnailPath: thumbnailExists ? thumbnailWebPath : undefined,
-    thumbnailUrl: thumbnailExists ? thumbnailWebPath : undefined,
-    duration: duration ? duration.toString() : undefined,
-    fileSize,
-    createdAt: new Date().toISOString(),
-    date: new Date().toISOString().split("T")[0].replace(/-/g, ""),
-    addedAt: new Date().toISOString(),
-  };
-
-  try {
-    const inserted = storageService.saveVideoIfAbsent(newVideo);
-    if (!inserted) {
-      cleanupUploadedVideo(file);
-      cleanupGeneratedThumbnail(validatedThumbnailPath);
-      const concurrentVideo = storageService.getVideoById(videoId);
-      if (concurrentVideo) {
-        return buildUploadDuplicateResult(file, concurrentVideo);
-      }
-
-      throw new Error(
-        `Concurrent upload conflict for ${videoId}: existing record could not be loaded`
-      );
-    }
-
-    return buildUploadSuccessResult(file, newVideo);
-  } catch (error) {
-    cleanupUploadedVideo(file);
-    cleanupGeneratedThumbnail(validatedThumbnailPath);
-    throw error;
-  }
-};
-
-// Helper responses
-const sendData = (res: Response, data: any) => {
-  res.status(200).json(data);
-};
-
-const sendSuccess = (res: Response, data: any, message: string) => {
-  res.status(200).json(successResponse(data, message));
-};
-
-// Extract language code from filename (e.g. "movie.en.vtt" -> "en")
-const getLanguageFromFilename = (filename: string): string | null => {
-  const parts = filename.split(".");
-  if (parts.length < 2) return null;
-  const langCode = parts[parts.length - 2];
-  if (/^[a-z]{2,3}(-[A-Z]{2})?$/i.test(langCode)) return langCode;
-  return null;
-};
-
-const resolveVideoWebPath = (absoluteVideoPath: string): string | null => {
-  const videosRoot = path.resolve(VIDEOS_DIR);
-  // nosemgrep: javascript.lang.security.audit.path-traversal.path-join-resolve-traversal.path-join-resolve-traversal
-  const normalizedPath = path.resolve(absoluteVideoPath);
-
-  if (
-    normalizedPath !== videosRoot &&
-    !normalizedPath.startsWith(`${videosRoot}${path.sep}`)
-  ) {
-    return null;
-  }
-
-  const relativePath = path.relative(videosRoot, normalizedPath);
-  if (!relativePath || relativePath.startsWith("..")) {
-    return null;
-  }
-
-  return `/videos/${relativePath.split(path.sep).join("/")}`;
-};
+// Re-export upload middleware/storage so existing `videoController.*` imports keep working.
+export {
+  upload,
+  uploadBatch,
+  uploadSubtitleMiddleware,
+  videoBatchUploadStorage,
+  videoUploadStorage,
+} from "./video/uploadMiddleware";
 
 /**
  * Get all videos
@@ -810,189 +324,6 @@ export const updateVideoDetails = async (
     success: true,
     video: updatedVideo,
   });
-};
-
-type ExistingVideoRecord = { id: string; channelUrl?: string };
-
-const BILIBILI_REQUEST_HEADERS = {
-  Referer: "https://www.bilibili.com",
-  "User-Agent":
-    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36",
-};
-
-const VIDEO_CONTENT_TYPE_BY_EXTENSION: Record<string, string> = {
-  ".mp4": "video/mp4",
-  ".webm": "video/webm",
-  ".mkv": "video/x-matroska",
-  ".avi": "video/x-msvideo",
-  ".mov": "video/quicktime",
-  ".m4v": "video/x-m4v",
-  ".flv": "video/x-flv",
-  ".3gp": "video/3gpp",
-};
-
-const getExistingVideoBySourceUrl = (sourceUrl: string): ExistingVideoRecord | null =>
-  (storageService.getVideoBySourceUrl(sourceUrl) as ExistingVideoRecord | null);
-
-const persistChannelUrlForVideo = (
-  existingVideo: ExistingVideoRecord | null,
-  channelUrl: string
-): void => {
-  if (!existingVideo) {
-    return;
-  }
-  storageService.updateVideo(existingVideo.id, { channelUrl });
-};
-
-const getBilibiliApiUrl = (videoId: string): string => {
-  if (videoId.startsWith("BV")) {
-    return `https://api.bilibili.com/x/web-interface/view?bvid=${videoId}`;
-  }
-  return `https://api.bilibili.com/x/web-interface/view?aid=${videoId.replace(
-    "av",
-    ""
-  )}`;
-};
-
-const fetchYouTubeChannelUrl = async (sourceUrl: string): Promise<string | null> => {
-  if (!isYouTubeUrl(sourceUrl)) {
-    return null;
-  }
-
-  const { executeYtDlpJson, getNetworkConfigFromUserConfig, getUserYtDlpConfig } =
-    await import("../utils/ytDlpUtils");
-  const userConfig = getUserYtDlpConfig(sourceUrl);
-  const networkConfig = getNetworkConfigFromUserConfig(userConfig);
-  const info = await executeYtDlpJson(sourceUrl, {
-    ...networkConfig,
-    noWarnings: true,
-  });
-
-  return info.channel_url || info.uploader_url || null;
-};
-
-const fetchBilibiliChannelUrl = async (
-  sourceUrl: string
-): Promise<string | null> => {
-  if (!isBilibiliUrl(sourceUrl)) {
-    return null;
-  }
-
-  const { extractBilibiliVideoId } = await import("../utils/helpers");
-  const videoId = extractBilibiliVideoId(sourceUrl);
-  if (!videoId) {
-    return null;
-  }
-
-  try {
-    const axios = (await import("axios")).default;
-    const response = await axios.get(getBilibiliApiUrl(videoId), {
-      headers: BILIBILI_REQUEST_HEADERS,
-    });
-
-    const ownerMid = response?.data?.data?.owner?.mid;
-    if (!ownerMid) {
-      return null;
-    }
-    return `https://space.bilibili.com/${ownerMid}`;
-  } catch (error) {
-    logger.error("Error fetching Bilibili video info:", error);
-    return null;
-  }
-};
-
-const fetchTwitchChannelUrl = async (
-  sourceUrl: string
-): Promise<string | null> => {
-  if (isTwitchChannelUrl(sourceUrl)) {
-    return normalizeTwitchChannelUrl(sourceUrl);
-  }
-
-  if (!isTwitchVideoUrl(sourceUrl)) {
-    return null;
-  }
-
-  const twitchVideoId = extractTwitchVideoId(sourceUrl);
-  if (!twitchVideoId) {
-    return null;
-  }
-
-  try {
-    const twitchVideo = await twitchApiService.getVideoById(twitchVideoId);
-    if (twitchVideo) {
-      return `https://www.twitch.tv/${twitchVideo.userLogin}`;
-    }
-  } catch (error) {
-    logger.error("Error fetching Twitch video info:", error);
-  }
-
-  const {
-    getChannelUrlFromVideo,
-    getNetworkConfigFromUserConfig,
-    getUserYtDlpConfig,
-  } = await import("../utils/ytDlpUtils");
-  const userConfig = getUserYtDlpConfig(sourceUrl);
-  const networkConfig = getNetworkConfigFromUserConfig(userConfig);
-  return getChannelUrlFromVideo(sourceUrl, networkConfig);
-};
-
-const resolveChannelUrl = async (sourceUrl: string): Promise<string | null> => {
-  const youtubeChannelUrl = await fetchYouTubeChannelUrl(sourceUrl);
-  if (youtubeChannelUrl) {
-    return youtubeChannelUrl;
-  }
-  const bilibiliChannelUrl = await fetchBilibiliChannelUrl(sourceUrl);
-  if (bilibiliChannelUrl) {
-    return bilibiliChannelUrl;
-  }
-  return fetchTwitchChannelUrl(sourceUrl);
-};
-
-const isMountVideoPath = (videoPath: string | undefined): boolean =>
-  typeof videoPath === "string" && videoPath.startsWith("mount:");
-
-const validateRawMountFilePath = (rawFilePath: string): void => {
-  if (!rawFilePath) {
-    throw new ValidationError("Invalid file path: empty or invalid", "videoPath");
-  }
-  if (rawFilePath.includes("..") || rawFilePath.includes("\0")) {
-    throw new ValidationError(
-      "Invalid file path: path traversal detected",
-      "videoPath"
-    );
-  }
-  if (!path.isAbsolute(rawFilePath)) {
-    throw new ValidationError("Invalid file path: must be absolute", "videoPath");
-  }
-};
-
-const resolveMountFilePath = (rawFilePath: string): string => {
-  validateRawMountFilePath(rawFilePath);
-  const filePath = normalizeSafeAbsolutePath(rawFilePath);
-  validateRawMountFilePath(filePath);
-  return filePath;
-};
-
-const assertMountFileExists = (filePath: string): void => {
-  if (!pathExistsTrustedSync(filePath)) {
-    throw new NotFoundError("Video file", filePath);
-  }
-  if (!statTrustedSync(filePath).isFile()) {
-    throw new ValidationError("Path is not a file", "videoPath");
-  }
-};
-
-const getVideoContentType = (filePath: string): string =>
-  VIDEO_CONTENT_TYPE_BY_EXTENSION[path.extname(filePath).toLowerCase()] ||
-  "video/mp4";
-
-const setMountVideoHeaders = (res: Response, filePath: string): void => {
-  res.setHeader("Access-Control-Allow-Origin", "*");
-  res.setHeader(
-    "Access-Control-Expose-Headers",
-    "Accept-Ranges, Content-Range, Content-Length"
-  );
-  res.setHeader("Content-Type", getVideoContentType(filePath));
 };
 
 /**
